[MRG] Ask for license agreement when run doc locally (#375)
* Ask for license agreement when run doc locally

* Also check if the dataset exists

* Update flake, test download data

* Move the download part to fetch function

* Fix print score of linear gaussian

* Rookie typos

* Download files to glm_data

* download dataset, fixing a few returns problem

* Minor fix of documentation

* Add what's new, overall fix in datasets.py

* Fix overindentation

* Clean up

* Fixing bug, makes CircleCI happy

* Fix function name

+++ pyglmnet/datasets.py
@@ -2,15 +2,13 @@
 A set of convenience functions to download datasets for illustrative examples
 """
 import os
+import os.path as op
 import sys
-import shutil
 import itertools
 import numpy as np
 from scipy.special import comb
 from urllib.request import urlretrieve
-from tempfile import TemporaryDirectory
 
-pbar = None
 _rgcs_license_text = """
 License
 -------
@@ -43,45 +41,73 @@ def _reporthook(count, block_size, total_size):
                      % (percent, progress_size / (1024 * 1024)))
 
 
-def fetch_tikhonov_data(dpath):
-    """
-    Downloads data for Tikhonov example and returns data frames
+def get_data_home(data_home=None):
+    """Return the path of the pyglmnet data dir.
+
+    Function from scikit-learn.
+
+    This folder is used by some large dataset loaders to avoid downloading the
+    data several times.
+
+    By default the data dir is set to a folder named 'glm_data' in the
+    user home folder.
 
     Parameters
     ----------
-    dpath: str
+    data_home : str | None
+        The path to pyglmnet data dir.
+    """
+    if data_home is None:
+        data_home = op.join('~', 'glm_data')
+    data_home = op.expanduser(data_home)
+    if not op.exists(data_home):
+        os.makedirs(data_home)
+    return data_home
+
+
+def fetch_tikhonov_data(dpath=None):
+    """Downloads data for Tikhonov example.
+
+    Parameters
+    ----------
+    dpath: str | None
         specifies path to which the data files should be downloaded.
+        default: None
 
     Returns
     -------
     dpath : str
-        The data path
+        The data path to Tikhonov dataset
     """
-
-    if os.path.exists(dpath):
-        shutil.rmtree(dpath)
-    os.mkdir(dpath)
-
-    base_url = "https://raw.githubusercontent.com/glm-tools/datasets/master"
+    dpath = get_data_home(data_home=dpath)
     fnames = ['fixations.csv', 'probes.csv', 'spiketimes.csv']
 
-    for fname in fnames:
-        url = base_url + "/tikhonov/" + fname
-        fname = os.path.join(dpath, fname)
-        urlretrieve(url, fname, _reporthook)
+    if not (op.isdir(dpath) and all(op.exists(op.join(dpath, fname))
+                                    for fname in fnames)):
+        base_url = (
+            "https://raw.githubusercontent.com/glm-tools/datasets/master"
+        )
+        fnames = ['fixations.csv', 'probes.csv', 'spiketimes.csv']
+
+        for fname in fnames:
+            url = base_url + "/tikhonov/" + fname
+            fname = os.path.join(dpath, fname)
+            urlretrieve(url, fname, _reporthook)
 
     return dpath
 
 
-def fetch_community_crime_data():
-    """
-    Downloads data for the community crime example,
-    removes missing values, extracts features, and
+def fetch_community_crime_data(dpath=None):
+    """Downloads community crime data.
+
+    This function removes missing values, extracts features, and
     returns numpy arrays
 
     Parameters
     ----------
-    None
+    dpath: str | None
+        specifies path to which the data files should be downloaded.
+        default: None
 
     Returns
     -------
@@ -96,15 +122,22 @@ def fetch_community_crime_data():
         raise ImportError('The pandas module is required for reading the '
                           'community crime dataset')
 
-    with TemporaryDirectory(prefix="tmp_glm-tools") as dpath:
-        fname = os.path.join(dpath, 'communities.csv')
-        base_url = ("http://archive.ics.uci.edu/ml/machine-learning-databases")
+    dpath = get_data_home(data_home=dpath)
+    file_name = 'communities.csv'
+
+    if not (op.isdir(dpath) and op.exists(op.join(dpath, file_name))):
+        fname = os.path.join(dpath, file_name)
+        base_url = (
+            "http://archive.ics.uci.edu/ml/machine-learning-databases"
+        )
         url = base_url + "/" + "communities/communities.data"
 
         urlretrieve(url, fname, _reporthook)
 
         # Read in the file
         df = pd.read_csv(fname, header=None)
+
+    df = pd.read_csv(op.join(dpath, file_name), header=None)
 
     # Remove missing values
     df.replace('?', np.nan, inplace=True)
@@ -119,63 +152,71 @@ def fetch_community_crime_data():
     return X, y
 
 
-def fetch_rgc_data(dpath, accept_rgcs_license=False):
-    """
-    Downloads data for spike trains prediction in retinal ganglia cells.
-    Please see https://github.com/glm-tools/datasets/RGCs/ for permission
-    to use the dataset
+def fetch_rgc_spike_trains(dpath=None, accept_rgcs_license=False):
+    """Downloads data for spike trains prediction in retinal ganglia cells.
 
     Parameters
     ----------
     dpath: str
         specifies path to which the data files should be downloaded.
+        default: None
     accept_rgcs_license: bool
         specify to true to accept the license to use the dataset
         default: False
 
     Returns
     -------
-    dpath : str
-        The data path
+    dpath : str | None
+        The data path for retinal ganglia cells dataset
+
+    Note
+    ----
+    See https://github.com/glm-tools/datasets/RGCs/ for permission
+    to use the dataset.
     """
-    if accept_rgcs_license:
-        answer = 'y'
-    else:
-        answer = input('%s\nAgree (y/[n])? ' % _rgcs_license_text)
-    if answer.lower() != 'y':
-        raise RuntimeError('You must agree to the license to use this '
-                           'dataset')
+    dpath = get_data_home(data_home=dpath)
+    file_name = 'data_RGCs.json'
 
-    if os.path.exists(dpath):
-        shutil.rmtree(dpath)
-    os.mkdir(dpath)
+    # if file already exist, read it from there
+    if not (op.isdir(dpath) and op.exists(op.join(dpath, file_name))):
+        # accept licence
+        if accept_rgcs_license:
+            answer = 'y'
+        else:
+            answer = input('%s\nAgree (y/[n])? ' % _rgcs_license_text)
+        if answer.lower() != 'y':
+            raise RuntimeError('You must agree to the license to use this '
+                               'dataset')
 
-    base_url = (
-        "https://raw.githubusercontent.com/glm-tools/datasets/master"
-    )
-    fnames = ['data_RGCs.json']
-
-    for fname in fnames:
-        url = base_url + "/RGCs/" + fname
-        fname = os.path.join(dpath, fname)
-        urlretrieve(url, fname, _reporthook)
+        base_url = (
+            "https://raw.githubusercontent.com/glm-tools/datasets/master"
+        )
+        fnames = ['data_RGCs.json']
+        for fname in fnames:
+            url = base_url + "/RGCs/" + fname
+            fname = os.path.join(dpath, fname)
+            urlretrieve(url, fname, _reporthook)
 
     return dpath
 
 
-def fetch_group_lasso_datasets():
-    """
-    Downloads and formats data needed for the group lasso example.
+def fetch_group_lasso_data(dpath=None):
+    """Downloads and formats data needed for the group lasso example.
 
-    Returns:
-    --------
-    design_matrix: pandas.DataFrame
-        pandas dataframe with formatted data and labels
+    Parameters
+    ----------
+    dpath: str | None
+        specifies path to which the data files should be downloaded.
 
-    groups: list
+    Returns
+    -------
+    X: numpy array, shape (n_samples, n_features)
+        The design matrix.
+    y: numpy array, shape (n_samples,)
+        The labels.
+    group: list
         list of group indicies, the value of the ith position in the list
         is the group number for the ith regression coefficient
-
     """
     try:
         import pandas as pd
@@ -254,28 +295,35 @@ def fetch_group_lasso_datasets():
 
         return feature_vector
 
-    positive_url = \
+    positive_url = (
         "http://hollywood.mit.edu/burgelab/maxent/ssdata/MEMset/train5_hs"
-    negative_url = \
+    )
+    negative_url = (
         "http://hollywood.mit.edu/burgelab/maxent/ssdata/MEMset/train0_5_hs"
+    )
 
-    with TemporaryDirectory(prefix="tmp_glm-tools") as dpath:
+    dpath = get_data_home(data_home=dpath)
+    fnames = ['pos', 'neg']
+    if not (op.isdir(dpath) and all(op.exists(op.join(dpath, fname))
+                                    for fname in fnames)):
+        pos_file = os.path.join(dpath, 'pos')
+        neg_file = os.path.join(dpath, 'neg')
+        urlretrieve(positive_url, pos_file, _reporthook)
+        urlretrieve(negative_url, neg_file, _reporthook)
+    else:
         pos_file = os.path.join(dpath, 'pos')
         neg_file = os.path.join(dpath, 'neg')
 
-        urlretrieve(positive_url, pos_file, _reporthook)
-        urlretrieve(negative_url, neg_file, _reporthook)
+    with open(pos_file) as posfp:
+        positive_sequences = [str(line.strip().upper()) for idx, line in
+                              enumerate(posfp.readlines())
+                              if ">" not in line and idx < 2 * 8000]
 
-        with open(pos_file) as posfp:
-            positive_sequences = [str(line.strip().upper()) for idx, line in
-                                  enumerate(posfp.readlines())
-                                  if ">" not in line and idx < 2 * 8000]
-
-        with open(neg_file) as negfp:
-            negative_sequences = [str(line.strip().upper()) for idx, line in
-                                  enumerate(negfp.readlines())
-                                  if ">" not in line and
-                                  idx < 2 * len(positive_sequences)]
+    with open(neg_file) as negfp:
+        negative_sequences = [str(line.strip().upper()) for idx, line in
+                              enumerate(negfp.readlines())
+                              if ">" not in line and
+                              idx < 2 * len(positive_sequences)]
 
     assert len(positive_sequences) == len(negative_sequences), \
         "lengths were not the same: p={pos} n={neg}" \
@@ -288,10 +336,11 @@ def fetch_group_lasso_datasets():
 
     df = pd.DataFrame(data=np.vstack((positive_vector_matrix,
                                       negative_vector_matrix)))
-    df.loc[0:positive_vector_matrix.shape[0], "Label"] = 1.0
-    df.loc[positive_vector_matrix.shape[0]:, "Label"] = 0.0
+    df.loc[0:len(positive_vector_matrix), "Label"] = 1.0
+    df.loc[len(positive_vector_matrix):, "Label"] = 0.0
 
-    design_matrix = df
-    groups = create_group_indicies_list()
+    X = df[df.columns.difference(["Label"])].values
+    y = df.loc[:, "Label"].values
+    group = create_group_indicies_list()
 
-    return design_matrix, groups
+    return X, y, group
